Make a few travis-related tweaks.
This makes a few changes to get the travis configuration working:

* Move `bootstrap` to the beginning of the list, so that new apt repos are
added immediately.

* Drop an `apt-get update` (since `bootstrap` does it anyway).

* Add back `rrdf` to `Suggests`, since (based on the git history) that was
only dropped to make the travis build happy. Note that I did **not** revert
any of the other diffs in that CL.

* Last, I noticed that `.travis.yml` was now giving a warning when doing the
CRAN check; I added it to `.Rbuildignore`, which involved removing
`.Rbuildignore` from `.gitignore`. If it's in there for other reasons, I'm
happy to revert.