force consistent type definition "symbol"
@hlapp Not sure if this is a complete fix for #171 / #172, but I noticed that the "symbol" attribute was being typed as an integer in the S4 definition of a state, but as character in the S4 definition of a uncertain state.

This change lets get_characters() run without error on the ontotrace-example.xml, but I haven't had a chance to dig deeper and see if it fixes all possible issues you flagged in #172.

(This is another nice feature of JSON over XML -- in XML the type of these attributes can be ambiguous, where as JSON has native atomic types for values).