Fixes broken handling of polymorphic and uncertain states
This fixes the handling of polymorphic and uncertain character states
for non-molecular (such as morphological) characters, as documented in
issues #171 and #172.

The change does not include code for allowing the user to specify a
custom symbol for polymorphic and uncertain states. The symbol in the
returned matrix will simply be the one given for the respective state
sets in the NeXML object.

Also includes a much expanded test.

Closes #171. Addresses much of #172 as well, except for giving the user
control over the symbol being used, and enabling queries as to which
cells are polymorphic.