Makes test suite more robust, including to missing packages (#181)
* Makes test suite more robust to missing packages

There are a number of tests that depend on packages in the suggested
category, and which are arguably needed by very few only. The changes
here will skip those tests rather than fail hard.

* Moves test cases that modify global object to the end

Also fixes test cases that were overriding the global object (which
happened to be inconsequential because that test case was at the end).

* Re-enables sparql test on Travis

It's not clear why the test was disabled - it does not seem time consuming.