Adds simplify option for table returned by get_metadata()
This won't make much of a difference (except for consolidating the
property/rel columns) if there are no nested metadata, but is quite
useful for tables that include nested metadata.

If set to `TRUE`, it might be useful to still consolidate the
type-specific LiteralMeta and ResourceMeta ID columns into a single
one. This is currently not done. However, IDs for meta elements are
rarely present (so usually these columns are all NAs), and the only
case something needs to be joined _to_ a metadata element (as opposed
to the other way) is when there _are_ nested meta elements.

Includes tests.