Enables parsing and serialization of XML-type literal metas
The value of LiteralMeta elements could be an XML literal. This is,
for example, used by Phenoscape to capture annotation of character
states with ontology terms by curators (using the Phenex application).
This change enables these values to be read and serialized, and to
be returned (converted to character strings) by `get_metadata()`.