Makes most S4 classes in the package pseudo-namespaced
In response to already encountered and possible future clashes of S4
class names defined here and elsewhere, this prefixes most of the
classes defined here with "nexml:". The only classes left unchanged
(for now?) are the ListOfxxxx classes, which have only minor roles as
glorified lists, and the "nexml" root class.

Arguably most if not all clients shouldn't be interacting directly with
classes other than the "nexml" class, and hence all others should arguably
be considered _internal_ to the package. Hence, how nice and short their
names are shouldn't matter to anyone outside the package, whereas short
and clean names ("tree", "meta", "char", "row", "seq", others anyone??)
do provide for a heightened risk of clashing with classes defined elsewhere.

Originally, the class hierarchy seems to have been designed with the desire
to have a 1:1 correspondence with the NeXML schema. I would argue that
makes the mistake of overloading and conflating the concern about mapping
properly to and from NeXML schema elements with properly naming S4 classes.
This change set solves this by introducing (also pseudo-namespaced for
easy removal in data extractions) slots `xmlName` and `idrefName`. For
the most part these have the same value for now, but there is no need for
them to do so now (unlike previously, where this was all tied to the S4 class
name).

Obviously, this change required changes to the test suite across the board
because the test suite does a lot of type testing. I tried not to change
anything else (for now).