Replaces the simmap_ex saved data file with source NeXML
The simmap examples and vignette loaded a saved nexml-simmap object with
unclear (really, unknown) provenance, and which would be certain to break
without a path to repair for any change to the S4 class structure or
definition. This replaces the saved data file with NeXML source data file
that should result in the equivalent object.

I obtained this file in the following way:
- Read in the simmap_ex saved data file at a codepoint with compatible
  class structure and definition.
- Wrote to NeXML using write.nexml().
- Corrected the erroneous LiteralMeta, property, and datatype values using
  the following perl one-liner:
    ```perl
    perl -p -e 'if (/simmap:reconstructions?/ || /simmap:stateChange/) { s/LiteralMeta/ResourceMeta/; s/property=/rel=/; s/ datatype=\"xsd:string\"//; }' < simmap_ex.xml
    ```
- Corrected the characters type to ContinuousCells. It was StandardCells,
  which caused the file to fail validation.
- Verified that the file now validates in full, using the online NeXML
  validator.