Fixes duplicate S4 class name issues without changing class names
It turns out that S4 classes defined with the same name but in different
packages are actually supported in R. Hence, pseudo-namespacing class
names as proposed in the #201 changeset should in principle not be
required. The mechanism for namespacing a class is in essence through
the package attribute of the class name string. The function packageSlot()
can be used to query and set the package attribute, and the function
className() generates a class name namespaced in this way from its two
arguments.

Therefore, this would allow us to keep the class names we want, not have
to worry about adding and stripping the pseudo-namespace to/from the
class name, and maintain full backward compatibility. All we would
have to do is use `className(name, "RNeXML")` instead of `name` when
new() is invoked.

However, that's unfortunately only the theory. In practice, two things
are at least currently broken, preventing this to be sufficient to make
it work:
* callNextMethod(), which is the key mechanism to recurse a generic down
  the inheritance chain, internally calls addNextMethod() in order to
  find and record the next method to be called in the chain. That
  function is passed all the information it would need (package names
  for namespacing super classes etc), but somehow it nonetheless
  errors out if at least the first argument in the signature is a class
  that's duplicated in another package.
* S4 generics dispatch, which is the key mechanism to find the appropriate
  method for a given set of arguments, somehow at some point loses the
  correct method for a signature in its lookup table if at least the first
  argument is of a class that's duplicated in another package.

This change set breaks down into three parts:
1. Create a convenience function New() that ensures the class name is
   properly namespaced. This solves the problem of making invocation of
   new() unambiguous.
2. Create a function that does the work of finding the next method that
   addNextMethod() should be doing instead, and caching the result in
   the calling environment in a way that callNextMethod() picks it up
   and then bypasses the invocation of addNextMethod().
3. Create a function that finds the right method for an S4 generic, in the
   way standardGeneric() should be doing but doesn't in certain cases.