Ensure NULL list elements in nexml slots are not converted to XML
This fixes an error in the simmap serialization/deserialization that
has manifested in the simmap vignette.

Obviously, this raises two questions. One, how did a NULL element
get into a nexml slot holding a list in the first place (shouldn't
result from XML parsing!). Two, why is this only surfacing in the
vignette and not caught in one of the simmap tests.

I don't know enough about the simmap data structure to write a
stricter test that would catch this (and potentially surface the
cause of a NULL element getting into the list in the first place).