Adds constructors for all classes in the hierarchy
This follows the initial example of using the New() convenience function
and the nexml.tree() constructor function that wraps that, and expands
it to all other classes in the nexml class hierarchy. These constructors
are now also exported so that users of the package can program with them
while being insulated from the exact inner workings of initializing and
naming classes.

The only classes not receiving a constructor function are the ListOfXXX
classes, which are simply glorified lists but not otherwise embedded in
the hierarchy (they also don't inherit from Base).

Also adds documentation, specifically here for the toplevel class (nexml).