Adds a utility function for expanding prefixed strings
This lays the basis for refactoring the way basic metadata are extracted.
At present the code either hardcodes the prefixes, which is clearly much
less than ideal, or converts to XML to then use XPATH queries, which wastes
a lot of time because it happens repeatedly for even only minor pieces of
metadata such as the license.