Fixes losing all but first list when concatenating ListOfMeta and meta
Revealed by writing stricter tests. unlist() does not take `...` as
arguments, so passing that on from the `c()` generic call will result
in all but the first element getting lost.

As an added benefit, `unlist()` by default already does the recursion
that a specially written function was doing, so can get rid of that
_and_ have a more robust and faster implementation :-)