Enables handling of files with rootedge
The NeXML schema provides for a `<rootedge/>` element for trees, see
[AbstractRootEdge]. Although the class hierarchy here included a RootEdge
class, it was not used, instantiated, or assigned in, resulting in the
rootedge element, if present, being in essence ignored and lost.

In practice, the RootEdge class is unnecessary, as a rootedge is an simply
and edge without a source. Since values for `source` and `target` slots
are not required here (unlike in the XML schema), the `edge` class can
serve for the rootedge slot, too.

The conversion to ape::phylo also ignored the rootedge slot, this is
(hopefully) fixed here too.

This change includes a number of tests, though a real life test file
seems hard to come by. The one included here was obtained by writing
the result of `rcoal(4)` to Newick, adding a rootedge in Newick, reading
it back in to ape::phylo, writing out to NeXML, and finally doctoring
the XML file to convert the root edge to a `<rootedge/>` element.
Apparently not even `rcoal()` will yield an actual root edge.

Closes #207.

[AbstractRootEdge]: http://www.nexml.org/doc/schema-1/trees/abstracttrees/#AbstractRootEdge