Makes metadata-per-block counting independent of proper names
The package methods for adding data elements often do not ensure proper
names are present when creating element lists, such as for example when
adding otus. This is in contrast to when these lists get created from
XML document content, because the conversion code uses lapply() universally.

Unfortunately, trying to catch and correct this at the constructor level
doesn't work for inexplicable reasons; it results in segfaults at some
point during the test suite. Perhaps the names property for list-containing
classes is read-only?